% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit.r
\name{clogit}
\alias{clogit}
\title{clogit}
\usage{
clogit(starts3, dat, otherdat, alts)
}
\arguments{
\item{starts3}{Starting values}

\item{dat}{Data matrix, see output from shiftSortX, alternatives with distance by column bind}

\item{otherdat}{Other data used in model}

\item{alts}{Number of alternative choices in model}
}
\description{
Conditional logit likelihood
}
\examples{
outputs:
ld - negative log likelihood
}
\keyword{fish}
