% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logit_input.r
\name{create_logit_input}
\alias{create_logit_input}
\title{create_logit_input}
\usage{
create_logit_input(modelInputData)
}
\arguments{
\item{modelInputData}{Data structure with information on variable choices and the alternative matrix (distances)}
}
\description{
Creates a data matrix that is consistent with the built in model forms
}
\examples{
outputs:
dataCompile - a data matrix
}
\keyword{fish}
