% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_startparams.r
\name{explore_startparams}
\alias{explore_startparams}
\title{explore_startparams}
\usage{
explore_startparams(space, startsr, dev, func, catch, choice, distance,
  otherdat)
}
\arguments{
\item{space}{Number of starting value permutations to test (the size of the space to explore)}

\item{startsr}{Average starting value parameters for revenue/location-specific covariates then cost/distance}

\item{dev}{How far to deviate from the average parameter values when exploring (random normal deviates)}

\item{func}{Name of likelihood function}

\item{catch}{Data corresponding to actual zonal choice}

\item{choice}{Data corresponding to actual catch}

\item{distance}{Data corresponding to distance}

\item{otherdat}{Other data (as a list)}
}
\value{
newstart - Chosen starting values with smallest likelihood \cr 
saveLLstarts - Likelihood values for each starting value permutation \cr 
savestarts - Starting value permuations (corresponding to each saved likelihood value) \cr
}
\description{
Subroutine to run find better starting values by exploring starting value parameter space
}
\examples{

}
