% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_func_true.r
\name{mc_func_true}
\alias{mc_func_true}
\title{Monte Carlo function}
\usage{
mc_func_true(
  mcnum,
  constp = c(0, 0, 0, 0),
  cdev,
  iinum,
  bwnum = 1,
  sertype = "gev",
  chodev = 1,
  alpha = 3,
  betac = -1,
  regconstant = 0
)
}
\arguments{
\item{mcnum}{MC iteration}

\item{constp}{Catch regression constant values}

\item{cdev}{Catch devation standard deviation}

\item{iinum}{Number of observations at each location}

\item{bwnum}{Bandwidth value for kernel}

\item{sertype}{Choice error type}

\item{chodev}{Choice error parameter}

\item{alpha}{Marginal utility from catch}

\item{betac}{Marginal disutility from distance}

\item{regconstant}{Catch regression constant exists}
}
\value{
Output from all models
}
\description{
MC function for Chen et al. 2022
}
