% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftSortX.r
\name{shiftSortX}
\alias{shiftSortX}
\title{shiftSortX}
\usage{
shiftSortX(x, ch, alts, ab, y)
}
\arguments{
\item{x}{Matrix of choice possibilites and distance from create_logit_input}

\item{ch}{Data corresponding to actual zonal choice}

\item{alts}{Number of alternative choices in model}

\item{ab}{Number of parameters in model? number of cost parameters + number of alts}

\item{y}{Data corresponding to actual catch scaled}
}
\description{
Shifts choices so that the chosen zone will be automatically the first one
}
\examples{
outputs:
d - matrix of choice possibilites and distance
}
\keyword{fish}
