% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_subroutine.r
\name{zonal_subroutine}
\alias{zonal_subroutine}
\title{zonal_subroutine}
\usage{
zonal_subroutine(modelInputData, dataCompile, betain, altcoeff, optimOpt,
  func, otherdat)
}
\arguments{
\item{modelInputData}{Information on model preference and parameters as LIST}

\item{dataCompile}{Data matrix from make_logit_input function as MATRIX}

\item{betain}{Initial parameter estimates for cost/distance}

\item{altcoeff}{Initial parameter estimates for revenue/location-specific covariates}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x]}

\item{func}{Name of likelihood function}
}
\description{
Subroutine to run chosen discrete choice model
}
\examples{
outputs:
OutLogit - [outmat1 se1 tEPM2] (coefs, ses, tstats)
clogitoutput - optimization information
seoumat2 - ses
MCM - Model Comparison metrics
H1 - inverse hessian
}
\keyword{fish}
