% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit.r
\name{clogit}
\alias{clogit}
\title{clogit}
\usage{
clogit(starts3, dat, alts, otherdat)
}
\arguments{
\item{starts3}{Starting values}

\item{alts}{Number of alternative choices in model}

\item{d}{Data matrix, see output from shiftSortX}
}
\description{
Conditional logit likelihood
}
\examples{
outputs:
ld - negative log likelihood
}
\keyword{fish}
