% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_subroutine.r
\name{zonal_subroutine}
\alias{zonal_subroutine}
\title{zonal_subroutine}
\usage{
zonal_subroutine(catch, choice, distance, otherdat, initparams, optimOpt,
  func)
}
\arguments{
\item{catch}{Data corresponding to actual zonal choice}

\item{choice}{Data corresponding to actual catch}

\item{distance}{Data corresponding to distance}

\item{otherdat}{Other data}

\item{initparams}{Initial parameter estimates for revenue/location-specific covariates then cost/distance}

\item{optimOpt}{Optimization options [max function evaluations, max iterations, (reltol) tolerance of x]}

\item{func}{Name of likelihood function}
}
\description{
Subroutine to run chosen discrete choice model
}
\examples{
outputs:
OutLogit - [outmat1 se1 tEPM2] (coefs, ses, tstats)
clogitoutput - optimization information
seoumat2 - ses
MCM - Model Comparison metrics
H1 - inverse hessian
}
\keyword{fish}
